SES-1056 Encode request using UTF-8.
Allows accessing remote endpoints with characters in the subquery that are not ASCII encodeable e.g.
Japanase text. This is needed to tell the remote server that the query/data is going to arrive in
utf-8 format if this is not done some servers like virtuoso 6 do not interpret the query string correctly.