Fix for build break introduced by commit 5823348.
In core/runtime, the POM's dependency scope for repository-contextaware
had been changed from 'compile' to 'runtime'. This didn't affect core,
which is why I didn't catch it. It did, however, break the compile
target for testsuites-sparql. I have reverted the dependency scope
to 'compile', and tested that 'mvn clean package' runs for the whole
repository.