SES-1669 : Push tuple and boolean handling methods up to new QueryResultHandler interface
This enables the removal of the previous generics arguments that were
used to distinguish between the two heirarchies. The new methods all now
throw QueryResultHandlerException.

For backwards compatibility, the old TupleQueryResultHandler methods
still throw TupleQueryResultHandlerException. As code is migrated to use
the new QueryResultWriter methods, it will be necessary for people to
start catching QueryResultHandlerException, and it should only be a
simple replace in catch and throws to move up to it, but it would be too
much of an issue for backwards compatibility at this point.

Existing implementations of TupleQueryResultWriter and
BooleanQueryResultWriter will now need to decide whether they can handle
the boolean or tuple methods, respectively, or whether they will throw
UnsupportedOperationException. By default all of the queryresultio-*
methods throw UnsupportedOperationException in these cases.