In this commit I introduced a common base class for all HTTP queries (i.e. queries of HttpRepository and SPARQLRepository). This base class is now already applied to SPARQLTupleQuery and HTTPTupleQuery.
The main difference of SPARQLTupleQuery and HTTPTupleQuery currently is
the handling of bindings: in the SPARQLRepository we replace them into
the queryString using regex, while in the HTTPTupleQuery we use the
Sesame proprietary protocol to pass them as request parameters. The next
steps here are obviously to define what we want to have.

Next steps will be to change also the other query types (graph +
boolean) to route them through the new HTTPClient and use common code.