This commit adds the implementation for routing also Graph and Boolean queries of both the SPARQLRepository and the HttpRepository through the HTTPClient, i.e. all query requests now share the same code basis.
For tuple and graph queries, the parsing is done in a background thread,
compare the respective BackgroundXXXResult classes.

The code of the Query implementation classes is now aligned so far, that
they are all derived from the AbstractHttpQuery. The main difference
between the HttpRepository and SPARQLRepository query implementation is
still that the SPARQLRepository versions apply some string modifications
(i.e. the bindings are inserted into the query), while the
HTTPRepository sends the bindings as GET parameters. See QueryStringUtil
for some details.

Some of the open tasks are marked as TODO in the code, e.g. there was
some handling of the baseURI for the SPARQL query version, which is not
yet dealt with in the new implementation. Except for the functionality
should have been moved to the respective HTTPClient method.

The most important advantage of the new unified implementation, is that
we do not have to take care for HTTP connections outside the HTTPClient.

Next steps and open tasks:

1) Route update queries through the HTTP Client as well
2) Remove apache http client from SPARQL repository
3) Further align HTTPRepository and SPARQLRepository, in particular make
HTTPRepository a subclass of SPARQLRepository (same for the connections
if possible)
4) Cosmetic cleanup and refactoring (e.g. in the query implementations
we always have the same kind of try catch block, which could be moved
into the HTTPClient itself (to return an instance of
QueryEvaluationException directly)