SES-1769 This commit adds the capability to limited the consumed memory by collections used to answer a SPARQL query. If the collections become to large a QueryEvaluationException is thrown.
The aim is to provide public sparql endpoints with more control on
resource usage.

Secondary effect is that its now easier to override the default
evaluation strategy to use custom container classes that are optimized
for the backing store.

The code is tested by having a new LimitedSizeNativeStoreConnection
that currently can only be configured in code.

The collections are only accessed via methods where ever possible,
for easier binary compatible support whenever a refactoring removes
the need for a collection.