SES-1720 : Add writer settings to allow backwards compatibility with RDF-1.0 where necessary
Both Plain Literals and Language Literals will be represented
differently internally from RDF-1.1.

Plain Literals will now always be typed and will always have xsd:string
as their datatype

Language Literals will now always be typed and will always have
rdf:langString as their datatype

The settings default to false, to allow backwards compatibility without
users having to switch it on everytime in the future.

These settings do not require any changes to the Literal interface,
other than to indicate in the contract that datatype will always be set
in future, and if and only if the datatype is rdf:langString, the
language will be set