SES-1720 : Add RDF-1.0 compatibility mode to SPARQL XML Writer
The workbench requires that literals not have datatypes in order for
them to be rendered correctly. This enables this backwards compatability
for the XSLT scripts in future.

The feature is off by default, so the rdf:langString and xsd:string will
still be written if they are available in other usecases