SES-1798 : Add specific settings into parsers
In the case of the Turtle parser, the verifyCharacter method always
fails now, and its name has changed to reflect that. There is no way to
use a helper method such as that and recover parsing reliably on a
tersely written Turtle document. NTriples can skip to the end of the
line and start again, but Turtle cannot rely on that.