SES-1814 : Move info.aduna.iteration package to sesame-util
Once it is in sesame-util it can be used by sesame-rio-api to add
Rio.write(Iteration...) methods in addition to Iterable that is
currently available. 

Not having Iteration there would mean that users would need to use
Iterations.asList() to convert their iteration to a static list before
using Rio.write(Iterable...).