SES-959 : Make Output URL optional to allow for non-deterministic tests
The test class currently uses a Base URL for parsing the openrdf
extension tests that is not deterministic. This makes it impossible to
generate an N-Triples file that requires the Base URL to parse a
relative URI in a document. The other relative URI tests rely on
xml:base, which is not present in the files for SES-679 and SES-959