SES-1844 : Add more tests for tuple results roundtripping
Some of the tests fail in the TSV parser, that seems to be a little lax
with the way it calculates positions so it does not smoothly roundtrip
for new line characters, tab characters, and bindings that start with
double-quotes and have no datatype. 

Might be useful to convert it to using a third-party TSV parser. For
now, the failing tests are commented out.