SES-1947 : Link stopAtFirstError to non-fatal errors
This restores the behaviour of the deprecated RDFParser.stopAtFirstError
function if the new functions are never used. If the new functions are
used, in particular, setNonFatalErrors(emptySet), then there may be a
discrepancy.

This commit is only aimed at restoring the previous behaviour for users
who are not migrating away from the deprecated functions from 2.7