SES-1972 : Call clear to reset bnode mappings in more parsers
Even though Rio parsers are not threadsafe, and are not designed to work
over multiple subsequent parses, if they incidentally work this way it
would be better to ensure, through the use of finally clauses, that
clear is called after each parse. This ensures that the bnode prefix is
reset each time.