Create extension points for RDFWriterTest
JSONLD, by default, does not preserve namespaces in JSONLDMode.EXPAND,
although RDFFormat.JSONLD rightly says that it is possible to preserve
namespaces (in a similar way to Turtle written using the N-Triples
subset will not preserve namespaces, but may be more efficient or
desirable in some circumstances). To be able to reuse this testsuite,
the SesameJSONLDWriterTest needs to let us know to use
JSONLDMode.COMPACT instead. This is enabled by offering an extension
point for both the WriterConfig and ParserConfig objects for the
relevant parsers, without otherwise changing the flow of the tests for
these cases.