SES-1776: add support for quads in SPARQLConnection getStatements
This change adds support for graph handling to the SPARQLConnection.
Implemented as an opt-in, the quad mode can be enabled in the
SPARQLRepository and is an immutable setting on newly created
SPARQLConnections (i.e. the current behavior will not be changed by
default).

The implementation makes use of a SELECT query pattern to also retrieve
the named graphs. 

Unit tests in SPARQLStoreConnectionTest show that both statements with a
named graph and statements from the default named graph can be
retrieved.

Note that this change does not resolve the issue with retrieving only
statements from the contexts provided as arguments to the method.