SES-2221: fix QueryPrologLexer to properly deal with comments
Due to a bug in the regular expression, the comment was not correctly
tokenized by the lexer. Before the change always the empty string was
returned as the first group.

The unit tests unfortunately did not catch this as they did not check
upon the "rest of query" string and not on the comment. 

This change provides a fix to the expression which properly matches the
entire comment line (including the #) and adds a bunch of unit tests.

Note: I could not get the regular expression working in a way that the
group matches the comment without the "#", hence I simply take the
substring in the readComment method.