SES-2242: fix basic authentication for SparqlRepository
When using the SparqlRepository against an endpoint that requires
authentication, the credentials were not passed as part of the request
(i.e. as a basic authentication header).

Investigating existing examples for HttpClient4 has shown that a
different type of HttpContext can be used. With this change examples in
our local environment (that also work with Sesame 2.7.x) work again on
Sesame 2.8.3