SES-2288: support easy configuration of SSL trust all (SPARQL Repo)
This change makes the managed http client configurable: users can
provide a custom HttpClientBuilder and set it to the SesameClientImpl.

In HttpClientBuilders a convenience method is provided to easily
construct a HttpClientBuilder for SSL Trust All purposes.

In a follow up for a major release the setHttpClientBuilder() method
will be exposed in the HttpClientDependent interface.