SES-860 slight refactoring to avoid further compilation problems
Eclipse does not pick up these compilation problems, but the Oracle JDK does - the issue is that due to type erasure the method signatures are ambiguous. Previously
fixed partially by changing the order of input parameters, but I decided that proper method rename is more clear and less likely to cause future problems.